backend/requirements.txt
Django>=5.1.6
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
Pillow
django-cors-headers
django-environ
django-extensions
black
